Unify virtual card edit/create modal
